pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "netflowsight"
version = "1.0.0"
description = "AI-Powered Network Traffic Analysis Platform - A fusion of high-performance PCAP parsing, threat detection, and intelligent reporting."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Prometheus Projects Team"}
]
keywords = ["pcap", "network-analysis", "threat-detection", "ai", "nfstream", "cybersecurity"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "nfstream>=6.5.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "fastmcp>=0.1.0",
    "pydantic>=2.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
netflowsight = "cli:main"

[project.urls]
"Homepage" = "https://github.com/prometheus-projects/NetflowSight"
"Bug Tracker" = "https://github.com/prometheus-projects/NetflowSight/issues"
"Documentation" = "https://github.com/prometheus-projects/NetflowSight/docs"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py39"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
